cleaned up the add_url in the DB
