+++ .sqlx/query-1538f8ced8c8fadf8da72dc615f0df5d9e96cc8e4ff9c8d9af88505438859e0b.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": " \n         SELECT user_id, user_name \n         FROM Users\n         ",
+  "query": "SELECT user_id, user_name FROM Users",
   "describe": {
     "columns": [
       {
@@ -22,5 +22,5 @@
       false
     ]
   },
-  "hash": "1fe4eb033e38d873d0a875093c4bfaea0f8d02c8a70744bc76cb067fee6e3a5c"
+  "hash": "1538f8ced8c8fadf8da72dc615f0df5d9e96cc8e4ff9c8d9af88505438859e0b"
 }
