+++ cli/src/add_url.rs
@@ -1,10 +1,8 @@
-// File: add_url.rs
 use color_eyre::Result;
-use db::{AddUrlOutcome, PgPool};
+use db::{add_url::AddUrlOutcome, PgPool};
 use uuid::Uuid;
 
 use crate::check_auth_status;
-
 pub async fn add_url(
     pool: &PgPool,
     url: &str,
@@ -12,5 +10,5 @@ pub async fn add_url(
     allow_existing: bool,
 ) -> Result<AddUrlOutcome> {
     check_auth_status(pool).await?;
-    db::add_url(pool, url, user_id, &allow_existing).await
+    db::add_url::add_url(pool, url, user_id, &allow_existing).await
 }
