+++ cli/src/auth.rs
@@ -1,3 +1,4 @@
+use db::user::get_username_by_id;
 use db::PgPool;
 use std::fs;
 use std::io::Read;
@@ -13,7 +14,7 @@ pub async fn check_auth_status(pool: &PgPool) -> color_eyre::Result<()> {
         file.read_to_string(&mut contents)?;
         let user_id = Uuid::parse_str(contents.trim())?;
 
-        match db::get_username_by_id(pool, user_id).await? {
+        match get_username_by_id(pool, user_id).await? {
             Some(username) => println!("Logged in with Username: {username}, User ID: {user_id}"),
             None => return Err(color_eyre::eyre::eyre!("User ID not found in database.")),
         }
