+++ cli/src/sign_up.rs
@@ -1,4 +1,4 @@
-use db::PgPool;
+use db::{user::create_user, PgPool};
 use std::path::Path;
 
 use crate::auth::{check_auth_status, persist_auth_session};
@@ -26,7 +26,7 @@ pub async fn sign_up(pool: &PgPool, username_opt: Option<String>) -> color_eyre:
 }
 
 pub async fn add_user(pool: &PgPool, username: &str) -> color_eyre::Result<()> {
-    let user_id = db::create_user(pool, username).await?;
+    let user_id = create_user(pool, username).await?;
     persist_auth_session(user_id)?;
 
     println!("User {username} added successfully with ID {user_id}!");
