+++ db/src/add_url.rs
@@ -1,0 +1,54 @@
+pub use sqlx;
+pub use sqlx::PgPool;
+use uuid::Uuid;
+
+#[derive(sqlx::FromRow, Debug)]
+pub struct Page {
+    pub page_id: Uuid,
+    pub user_id: Uuid,
+    pub url: String,
+}
+
+#[allow(clippy::module_name_repetitions)]
+#[derive(Debug)]
+pub enum AddUrlOutcome {
+    Created(Page),
+    Existing(Page),
+}
+
+pub async fn add_url(
+    pool: &PgPool,
+    url: &str,
+    user_id: Uuid,
+    allow_existing: &bool,
+) -> color_eyre::Result<AddUrlOutcome> {
+    let new_page_id = Uuid::new_v4();
+
+    let upsert_result = sqlx::query_as!(
+        Page,
+        "INSERT INTO Pages (page_id, user_id, url) VALUES ($1, $2, $3)
+         ON CONFLICT (user_id, url) DO UPDATE SET url = EXCLUDED.url
+         RETURNING *",
+        new_page_id,
+        user_id,
+        url
+    )
+    .fetch_one(pool)
+    .await?;
+
+    if upsert_result.page_id == new_page_id {
+        println!("URL '{}' added successfully.", upsert_result.url);
+        Ok(AddUrlOutcome::Created(upsert_result))
+    } else if *allow_existing {
+        println!(
+            "URL '{}' exists but re-adding is allowed.",
+            upsert_result.url
+        );
+        Ok(AddUrlOutcome::Existing(upsert_result))
+    } else {
+        Err(color_eyre::eyre::eyre!(
+            "URL '{}' already exists and re-adding is not allowed.",
+            upsert_result.url
+        ))
+    }
+}
